% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{check_figs}
\alias{check_figs}
\title{Check figs}
\usage{
check_figs(
  figsDir = "figs",
  scripts = list.files(pattern = "\\\\.R$"),
  notebooks = list.files(pattern = "\\\\.Rmd$"),
  fig_exts = c("pdf", "png", "jpg")
)
}
\arguments{
\item{figsDir}{Figures directory. Use empty}

\item{scripts}{character. File names of scripts to check}

\item{notebooks}{character. File names of Notebooks to check}

\item{fig_exts}{character. Extensions associated to figures, e.g. 'pdf',
'jpg'. Do not include a "." before the extension.}
}
\value{
data.table
}
\description{
Look for matches figures in both .R scripts and .Rmd notebooks
}
