% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggAngleXLabs}
\alias{ggAngleXLabs}
\title{Angle ggplot x labels}
\usage{
ggAngleXLabs(angle = 90)
}
\arguments{
\item{angle}{Angle used to tilt x labels (default = 90)}
}
\value{
ggplot
}
\description{
Angle ggplot x labels
}
