% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggBoxplot}
\alias{ggBoxplot}
\title{ggBoxplot}
\usage{
ggBoxplot(xFactor, yNumeric, outLCol = NA, outLAlpha = 0.5)
}
\arguments{
\item{xFactor}{factor. Factor vector specifying the groups}

\item{yNumeric}{numeric. Numeric vector specifying the values per observation}

\item{outLCol}{character. Color of outliers. NA for hidding outliers.}

\item{outLAlpha}{numeric. Value from 0 to 1 to set transparency of outliers.}
}
\value{
ggplot
}
\description{
ggplot2 boxplot shortcut
}
\examples{
ggBoxplot(iris$Species, iris$Petal.Length)
}
