% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{intervalHeatmap}
\alias{intervalHeatmap}
\title{Interval heatmap}
\usage{
intervalHeatmap(
  x,
  y,
  z,
  x_n,
  y_n,
  nCores = 1,
  dig.lab = 1,
  outputMatrix = FALSE,
  dispNum = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{x}{numeric. X dimension}

\item{y}{numeric. Y dimension}

\item{z}{logical or numeric. Z dimension to be summarized}

\item{x_n}{numeric. Number of groups for x dimension to be cut into}

\item{y_n}{numeric. Number of groups for y dimension to be cut into}

\item{nCores}{numeric. Number of cores to use to process data}

\item{dig.lab}{numeric. Number of decimal positions to place in labels}

\item{outputMatrix}{logical. Set to TRUE to output instead the matrix}

\item{dispNum}{logical. Set to FALSE for not showing the numeric labels at
each cell}

\item{na.rm}{logical. Indicating if NA values should be removed to calculate
mean Z}
}
\value{
matrix or
}
\description{
Plots the mean value of a value Z when divided into groups defined by
intervals at variables X and Y
}
\examples{
intervalHeatmap(x = iris$Sepal.Length, y = iris$Sepal.Width,
z = as.numeric(iris$Species), x_n = 5, y_n = 5)
}
