% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\docType{package}
\name{m.shortcuts}
\alias{m.shortcuts}
\alias{_PACKAGE}
\alias{m.shortcuts-package}
\title{m.shortcuts: A personal package to make my life easier}
\description{
The m.shortcuts package has a disparage collection of functions used in
common bioinformatics pipelines, and work from
\href{https://angelcampos.github.io/}{Miguel Angel Garcia-Campos}.
}
\author{
\strong{Maintainer}: Miguel Angel Garcia-Campos \email{angel.campos@weizmann.ac.il} (\href{https://orcid.org/0000-0001-5223-8174}{ORCID})

}
\keyword{ggplot2}
