% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot for ggplot2}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot2. Input plot objects}

\item{plotlist}{list. (optional) Input ggplots as a list}

\item{cols}{numeric. Number of columns to distribute plots}

\item{layout}{matrix. Optional layout to order plots}
}
\value{
NULL. plots input ggplots
}
\description{
Multiplot for ggplot2 by Winston Chang from
\href{http://www.cookbook-r.com/}{Cookbook for R}
}
\author{
Winston Chang
}
