% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{rmTmp}
\alias{rmTmp}
\title{Remove tmp objects from global environment}
\usage{
rmTmp(collectGarbage = TRUE)
}
\arguments{
\item{collectGarbage}{logical. Activate garbage collector}
}
\value{
matrix By default outputs the gc() output matrix
}
\description{
Remove tmp objects from global environment
}
