% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{lsos}
\alias{lsos}
\title{Improved list of objects}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{n}{numeric. Number of results to show when setting head to TRUE}

\item{pos}{numeric. an alternative argument to name for specifying the environment as
a position in the search list. Mostly there for back compatibility.}

\item{pattern}{character. pattern to match objects listed}

\item{order.by}{character. Name of column to order results by}

\item{decreasing}{logical. Set to FALSE if increasing order is desired}

\item{head}{logical. By default limits results to n results}
}
\value{

}
\description{
Prints a table showing the size of biggest objects in the global environment
(by default). Based on code by Dirk Eddelbuettel at
\href{https://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}{StackOverflow}
}
\author{
Dirk Eddelbuettel https://stackoverflow.com/users/143305/dirk-eddelbuettel
}
